Settings Added, Currency symbol is now changeable
