+++ src/core/Plugin/Plugin.css
@@ -23,7 +23,7 @@
 
 .sbs-budget .table > span {
     display: flex;
-    padding: 0.5rem 1rem;
+    padding: 0.75rem 1rem;
     font-size: var(--font-adaptive-small);
     align-items: center;
 }
