+++ src/core/Plugin/SettingTab/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingTab';
