+++ src/core/Plugin/index.ts
@@ -1,0 +1,1 @@
+export { Plugin } from './Plugin';
