+++ src/core/contexts.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { Settings } from './types';
+
+export const SettingsContext = createContext<Settings | undefined>(undefined);
