+++ src/ui/Value/Value.tsx
@@ -1,10 +1,13 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { SettingsContext } from '../../core/contexts';
 
 export interface ValueProps {
     value: number,
 }
 
 export const Value = ({ value }: ValueProps) => {
+    const { currency } = useContext(SettingsContext);
+
     const roundValue = Math.round(value * 100) / 100;
     const normalizedValue = Math.abs(roundValue);
 
@@ -21,7 +24,7 @@ export const Value = ({ value }: ValueProps) => {
 
     return (
         <span className={`dir-${dir}`}>
-            { `${ico}€${normalizedValue.toFixed(2)}` }
+            {`${ico}${currency}${normalizedValue.toFixed(2)}`}
         </span>
     );
 };
